+++ src/util.rs
@@ -4,6 +4,6 @@ pub mod instrument;
 pub mod order_status;
 pub mod time;
 
-pub use instrument::InstrumentInfo;
+pub use instrument::{InstrumentInfo, InstrumentInfoError};
 pub use order_status::OrderStatus;
 pub use time::{rithmic_to_unix_nanos, rithmic_to_unix_nanos_precise};